+++ packages/web_benchmarks/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `flutter_frame.total_time`, `flutter_frame.build_time`, and `flutter_frame.raster_time`
 metrics to benchmark results. These values are derived from the Flutter `FrameTiming` API.
+* Expose a new library `metrics.dart` that contains definitions for the benchmark metrics.
 ## 3.0.0
 
 * **Breaking change:** removed the `initialPage` parameter from the `serveWebBenchmark`
