+++ packages/web_benchmarks/lib/src/recorder.dart
@@ -18,6 +18,7 @@ import 'package:meta/meta.dart';
 import 'package:web/web.dart' as html;
 
 import 'common.dart';
+import 'metrics.dart';
 
 /// The number of samples from warm-up iterations.
 ///
